#130 - Fix service usage in Rails.
In Rails 2.3, helpers could access controller methods, and so, services
instantiated in application_controller were used by helpers.

In Rails 4, these are separated. So, helpers cannot access them.
Services module was created for this. It was included from both
controllers and helpers. This means, there are multiple references to
the same service, and are accessed separately from controllers and
helpers.

However, this causes issues during stubbing in RSpec. Especially, with
render_views option. In this case, controller specs render views as
well. So, the services used cannot be stubbed easily (they're called
from controllers sometimes, from helpers sometimes).

So, services are accessed using ServiceCache now, for production. For
specs, there is a way to stub those services (using TestServiceCache).
The ServiceCacheStrategy is used to decide between the two caches. The
new Rails configuration property (java_services_cache) is used by the
ServiceCacheStrategy. It is different in test.rb and production.rb.

In most cases, stub_service is all you need. You don't need to worry
about it.