Fixing failing rspec because of oauth2_provider engine
Because of the inclusion of oauth2_provider engine, all navigation routes had to be prepended with either 'main_app' or 'oauth_engine'. This caused few specs to fail because of the way urls were stubbed before. Changed it by including a EngineUrlHelper which stubs all header links and oauth2_provider links.