#000 - Fix inline ERB partials again.
This reverts a part of 8916c9efbfc30a2868b666aa558c5a26c1209d13. The part which
handles '<%== render_json(:partial => ...)'. This is because <%== in ERB means
it does not need to be escaped, etc. Needs more thought before doing that.

Also, add an error message, when parsing of locals and scope fails, in
one specific way. It used to fail silently till now.