#795 - Move cctray init to end.
CCTray initialization was conflicting with pipelineSqlMapDao initialization, because pipelineSqlMapDao
was loading from DB, into cache, without synchronization. Since cctray initialize is asynchronous, it
finished, and tried to use the cache, while it was being loaded, causing a concurrent modification
exception. So, this change moves the cctray init to the end of the application initializers.