#496 - Ability to have short-lived sessions for API requests.
When an API request starts a session, the idle time of that session is set to (a configurable value,
with default of) 300 seconds. This makes sure that sessions created due to API requests (usually
from scripts) time out, even if they do not send back the JSESSIONID cookie that the Go Server
sets, during every login.

Since this timeout affects only sessions started from API requests, it does not affect normal
functioning of the Go Server. Moreover, if a script does use the cookie and return it to the Go
Server, it continues to work, since the session will not be considered idle.