#000 - Trying to fix flaky UserSqlMapDaoCachingTest
changed assertion for one of the tests so that the values are not hardcoded
A bunch of tests used to fail because of thread-level timeout happening in enabledUserCacheShouldBeThreadSafe.
For now increased the timeout so the tests pass. Added a stop watch to record the operation times to see which operation takes long.