Fix broken tests because of classpath fudging.
A few commits earlier, the server's pom.xml had an ant-run plugin that
would spit out the [test
classpath](https://github.com/gocd/gocd/blob/04f352ac4cce4425c1d330025db90f8641006bc1/server/pom.xml#L758-L761)
to a file that was then picked up by rake to run rspecs
[server/run_rspec_tests.rake](https://github.com/gocd/gocd/blob/04f352ac4cce4425c1d330025db90f8641006bc1/server/run_rspec_tests.rake#L16-L23).

This was in commit 94db8a4, but caused build failures because we
overlooked a couple of entries in the classpath.

This commit fixes some of those issues in addition to reducing some
redundancies in the tlb configuration introduced because of cobertura.