#1133 extracted interface under CachedFileGoConfig and MergeGoConfig
Expectations towards MergedGoConfig and CachedFileGoConfig are the same,
added CachedGoConfig interface as a specification of that.
Now we should think in terms that there are 2 ways to deliver a cached
config service:
- by caching a simple xml file. When reading, it just returns object
from memory. When writing it saves xml to disk and updates memory
object. (like before)
- by caching latest merged config from file and config repos. When
reading it returns latest merged config instance from memory. When
writing it routes all changes to file.

Dao has been updated so that it actually depends on CachedGoConfig
interface.

CachedGoConfigBaseTest is a set of unit tests to execute against any
implementation of CachedGoConfig.

CachedFileGoConfigTest effectively runs the same tests that old
CachedGoConfigTest did.