#1133 protect against scheduling pipeline with old configuration revision
By analogy to assertMaterialsMatch. So far user could be the reason to
break materials consistency.
Now there is possibility that pipeline config has older revision than
what is proposed in build cause. So we must throw up in this case as
well.
This is a low level, late protection. So perhaps more should be done
higher so that this does not happen too often.