#1133 pipeline will not schedule if part is invalid
added test which asserts that when pipeline material and config repo are
the same and configuration part from repo cannot be parsed then pipeline
will not schedule.
This already passes without any changes because revisions config and
material revisions must match, while invalid part is not imported to
merged config.