#1210 - Changed configuration implementation based on feedback
ssl configuration will have to be setup using system properties and not config file
Default are the same as earlier implementation
included/excluded ciphers and protocols can be configured using a comma separated list.

Context: The plan is to use a common config file for all system properties, instead of multiple config files.
This will be done as separate story (#1313)