Temporarily "Fix" rpm build issue
If redhat-rpm-config package is installed, it creates a file called /usr/lib/rpm/redhat/macros, which changes
the %__spec_install_pre variable in the RPM spec, so that it deletes RPM_BUILD_ROOT (if it is not "/"). This
causes all the carefully setup installer files in dist/linux_package (which is the RPM_BUILD_ROOT) to be
deleted, failing the build.