Handling eviction/expiry of composite-keys from GoCache
When a composite key is saved to cache, there are two entries created in the cache. One for the composite-key itself containing the actual value. Another entry to keep track of all sub-keys for a given parent. When a parent cache entry is explicitly removed, the corresponding caches for subkeys are removed as well. When a subkey is explicitly removed, the sub-key list with the parent cache entry is updated as well. However, when a sub-key entry gets evicted from the cache when it gets full or when it expires, the parent cache entry wasn't updated. This commit fixes that.