#1133 fix hasSamePipelinesAs implementation
In EnvironmentsConfig any EnvironmentConfig instance must have unique
pipelines which are final members of that environment. Regardless of
config sources of that environment.
So in case when
EnvironmentsConfig.validateNotADuplicate(EnvironmentConfig
environmentConfig) is called on environment with many sources, we should
still validate that MergeEnvironmentConfig is the only definition for
that environment.